Apply BigInt changes

+++ libs/sdk-flutter/lib/breez_sdk.dart
@@ -593,7 +593,7 @@ extension SDKConfig on Config {
     String? defaultLspId,
     String? apiKey,
     double? maxfeePercent,
-    int? exemptfeeMsat,
+    BigInt? exemptfeeMsat,
     NodeConfig? nodeConfig,
   }) {
     return Config(
